Add Plan 9 support
